Download Firebase iOS SDK from Google instead of through CocoaPods
CocoaPods seems to contain a different version of the Firebase iOS SDK
than exists on CocoaPods. The CocoaPods version is missing static
libraries and frameworks that are required to link. The official
solution for working with CocoaPods is to use Xcode’s binary copy
feature, but we don’t want to have to do this step for working with Qt
projects.